+++ src/aui-char-counter/js/aui-char-counter.js
@@ -5,7 +5,8 @@
  */
 
 var L = A.Lang,
-    isNumber = L.isNumber;
+    isNumber = L.isNumber,
+    isString = L.isString;
 
     A.Node.DOM_EVENTS.compositionend = 1;
     A.Node.DOM_EVENTS.compositionstart = 1;
@@ -46,6 +47,17 @@ var CharCounter = A.Component.create({
      */
     ATTRS: {
         /**
+         * ARIA atomic attribute that describes assistive technologies will present all, or only parts of, the changed region based on the change notifications defined by the aria-relevant attribute.
+         *
+         * @attribute atomic
+         * @default true
+         * @type {Boolean}
+         */
+        atomic: {
+            value: true
+        },
+
+        /**
          * Node or Selector to display the information of the counter.
          *
          * @attribute counter
@@ -57,6 +69,17 @@ var CharCounter = A.Component.create({
         },
 
         /**
+         * ARIA describedby attribute that describes the current element.
+         *
+         * @attribute describedby
+         * @default ''
+         * @type {String}
+         */
+        describedby: {
+            value: ''
+        },
+
+        /**
          * Node or Selector for the input field. Required.
          *
          * @attribute input
@@ -65,6 +88,19 @@ var CharCounter = A.Component.create({
          */
         input: {
             setter: A.one
+        },
+
+        /**
+         * ARIA live attribute to help assistive technology properly read updates
+         * to the number of characters remaining.
+         *
+         * @attribute live
+         * @default 'polite'
+         * @type {String}
+         */
+        live: {
+            validator: isString,
+            value: 'polite'
         },
 
         /**
@@ -82,6 +118,20 @@ var CharCounter = A.Component.create({
             },
             validator: isNumber,
             value: Infinity
+        },
+
+        /**
+         * Boolean indicating if use of the WAI-ARIA Roles and States
+         * should be enabled.
+         *
+         * @attribute useARIA
+         * @default true
+         * @type Boolean
+         */
+        useARIA: {
+            value: true,
+            validator: L.isBoolean,
+            writeOnce: 'initOnly'
         }
     },
 
@@ -138,6 +188,7 @@ var CharCounter = A.Component.create({
          */
         bindUI: function() {
             var instance = this;
+
             var input = instance.get('input');
 
             instance.publish('maxLength');
@@ -162,7 +213,9 @@ var CharCounter = A.Component.create({
          */
         syncUI: function() {
             var instance = this;
+
             var counter = instance.get('counter');
+            var useAria = instance.get('useARIA');
 
             if (counter) {
                 var value = instance.get('input').val();
@@ -170,6 +223,10 @@ var CharCounter = A.Component.create({
                 var counterValue = instance.get('maxLength') - instance._getNormalizedLength(value);
 
                 counter.html(counterValue);
+            }
+
+            if (useAria) {
+                this._syncAriaControlsUI();
             }
         },
 
@@ -199,6 +256,7 @@ var CharCounter = A.Component.create({
          */
         checkLength: function() {
             var instance = this;
+
             var input = instance.get('input');
 
             var returnValue = false;
@@ -311,6 +369,7 @@ var CharCounter = A.Component.create({
          */
         _setMaxLength: function(v) {
             var instance = this;
+
             var input = instance.get('input');
 
             if (input && (v < Infinity)) {
@@ -318,6 +377,50 @@ var CharCounter = A.Component.create({
             }
 
             return v;
+        },
+
+         /**
+         * Updates the aria attribute for the component.
+         *
+         * @method _syncAriaControlsUI
+         * @protected
+         */
+        _syncAriaControlsUI: function() {
+            var instance = this;
+
+            instance.plug(
+                A.Plugin.Aria,
+                {
+                    attributes: {
+                        describedby: 'describedby',
+                    },
+                    attributeNode: instance.get('input')
+                }
+            );
+
+            var describedBy = instance.get('describedby');
+
+            describedBy = A.one('#' + describedBy);
+
+            if (describedBy) {
+                var atomic = instance.get('atomic');
+                var live = instance.get('live');
+
+                this.aria.setAttributes(
+                    [
+                        {
+                            name: 'atomic',
+                            node: describedBy,
+                            value: atomic
+                        },
+                        {
+                            name: 'live',
+                            node: describedBy,
+                            value: live
+                        }
+                    ]
+                );
+            }
         }
     }
 });