+++ src/aui-surface/js/aui-surface-app.js
@@ -694,10 +694,24 @@ A.SurfaceApp = A.Base.create('surface-app', A.Base, [], {
      * @private
      */
     _syncScrollPosition: function(opt_replaceHistory) {
-        win.scrollTo(
-            opt_replaceHistory ? this.pageXOffset : 0,
-            opt_replaceHistory ? this.pageYOffset : 0
-        );
+        var hash = win.location.hash,
+            hashEl,
+            pos = [0, 0];
+
+        if (opt_replaceHistory) {
+            win.scrollTo(this.pageXOffset, this.pageYOffset);
+            return;
+        }
+
+        if (hash) {
+            hashEl = A.one(hash);
+        }
+
+        if (hashEl) {
+            pos = hashEl.getXY();
+        }
+
+        win.scrollTo(pos[0], pos[1]);
     },
 
     /**
@@ -739,15 +753,16 @@ A.SurfaceApp = A.Base.create('surface-app', A.Base, [], {
     ATTRS: {
         /**
          * Defines the default document title in case the screen doesn't have
-         * any `title`.
+         * any `title`. If null, the default browser behavior for setting the
+         * title is used for when there's no title HTML tag on the document.
          *
          * @attribute defaultTitle
-         * @default Home
-         * @type {String}
+         * @default null
+         * @type {String|null}
          */
         defaultTitle: {
-            validator: A.Lang.isString,
-            value: 'Home'
+            validator: A.Lang.isString || A.Lang.isNull,
+            value: null
         },
 
         /**