+++ src/aui-toggler/js/aui-toggler-delegate.js
@@ -110,7 +110,8 @@ var TogglerDelegate = A.Component.create({
          */
         expanded: {
             validator: isBoolean,
-            value: true
+            value: true,
+            writeOnce: true
         },
 
         /**
@@ -190,9 +191,7 @@ var TogglerDelegate = A.Component.create({
         renderUI: function() {
             var instance = this;
 
-            if (instance.get('closeAllOnExpand')) {
-                instance.createAll();
-            }
+            instance._createAll();
         },
 
         /**
@@ -203,12 +202,21 @@ var TogglerDelegate = A.Component.create({
          */
         bindUI: function() {
             var instance = this;
+
             var container = instance.get('container');
             var header = instance.get('header');
+            var toggleEvent = instance.get('toggleEvent');
 
             instance._eventHandles = [
-                container.delegate([instance.get('toggleEvent'), 'keydown'], A.bind('headerEventHandler', instance), header),
-                instance.on('toggler:animatingChange', A.bind('_onAnimatingChange', instance))
+                container.delegate(
+                    [toggleEvent, 'keydown'],
+                    A.bind('headerEventHandler', instance),
+                    header
+                ),
+                instance.on(
+                    'toggler:animatingChange',
+                    A.bind('_onAnimatingChange', instance)
+                )
             ];
         },
 
@@ -221,9 +229,12 @@ var TogglerDelegate = A.Component.create({
         destructor: function() {
             var instance = this;
 
-            AArray.each(instance.items, function(item) {
-                item.destroy();
-            });
+            AArray.each(
+                instance.items,
+                function(item) {
+                    item.destroy();
+                }
+            );
 
             instance.items = null;
 
@@ -238,7 +249,7 @@ var TogglerDelegate = A.Component.create({
         collapseAll: function(payload) {
             var instance = this;
 
-            instance.createAll();
+            instance._createAll();
 
             A.Array.invoke(instance.items, 'collapse', payload);
         },
@@ -246,16 +257,27 @@ var TogglerDelegate = A.Component.create({
         /**
          * Forces toggler creation on delegated header elements.
          *
-         * @method createAll
+         * @method _createAll
+         * @protected
          */
-        createAll: function() {
+        _createAll: function() {
             var instance = this;
 
-            instance.get('container').all(instance.get('header')).each(function(header) {
-                if (!header.getData('toggler')) {
-                    instance._create(header);
+            var container = instance.get('container');
+            var closeAllOnExpand = instance.get('closeAllOnExpand');
+            var expanded = instance.get('expanded');
+            var header = instance.get('header');
+
+            container.all(header).each(
+                function(header) {
+                    var markupCollapsed = header.hasClass(CSS_TOGGLER_HEADER_COLLAPSED);
+                    var markupExpanded = header.hasClass(CSS_TOGGLER_HEADER_EXPANDED);
+
+                    if ((closeAllOnExpand && !markupCollapsed) || (expanded && markupCollapsed) || (!expanded && !markupExpanded && !markupCollapsed) && !header.getData('toggler')) {
+                        instance._create(header);
+                    }
                 }
-            });
+            );
         },
 
         /**
@@ -266,7 +288,7 @@ var TogglerDelegate = A.Component.create({
         expandAll: function(payload) {
             var instance = this;
 
-            instance.createAll();
+            instance._createAll();
 
             A.Array.invoke(instance.items, 'expand', payload);
         },
@@ -279,6 +301,7 @@ var TogglerDelegate = A.Component.create({
          */
         findContentNode: function(header) {
             var instance = this;
+
             var content = instance.get('content');
 
             var contentNode = header.next(content) || header.one(content);
@@ -303,32 +326,39 @@ var TogglerDelegate = A.Component.create({
         headerEventHandler: function(event) {
             var instance = this;
 
+            var closeAllOnExpand = instance.get('closeAllOnExpand');
+            var content = instance.get('content');
+
             if (instance.animating) {
                 return false;
             }
 
             var target = event.currentTarget;
+
             var toggler = target.getData('toggler') || instance._create(target);
 
-            if (Toggler.headerEventHandler(event, toggler) && instance.get('closeAllOnExpand')) {
-                var wrappingContent = toggler.get('content').ancestor(instance.get('content'));
+            if (Toggler.headerEventHandler(event, toggler) && closeAllOnExpand) {
+                var wrappingContent = toggler.get('content').ancestor(content);
 
-                AArray.each(instance.items, function(item) {
-                    if ((item !== toggler) && item.get('expanded')) {
+                AArray.each(
+                    instance.items,
+                    function(item) {
+                        if ((item !== toggler) && item.get('expanded')) {
 
-                        if (wrappingContent) {
-                            var itemContent = item.get('content');
+                            if (wrappingContent) {
+                                var itemContent = item.get('content');
 
-                            if ((itemContent !== wrappingContent) && wrappingContent.contains(itemContent)) {
+                                if ((itemContent !== wrappingContent) && wrappingContent.contains(itemContent)) {
+                                    item.collapse();
+                                }
+                            }
+                            else {
                                 item.collapse();
                             }
-                        }
-                        else {
-                            item.collapse();
-                        }
 
+                        }
                     }
-                });
+                );
             }
         },
 
@@ -340,8 +370,12 @@ var TogglerDelegate = A.Component.create({
          * @protected
          */
         _create: function(header) {
-            var instance = this,
-                expanded = instance.get('expanded');
+            var instance = this;
+
+            var animated = instance.get('animated');
+            var expanded = instance.get('expanded');
+            var toggleEvent = instance.get('toggleEvent');
+            var transition = instance.get('transition');
 
             // Prioritize markup information to decide whether it's expanded or
             // not
@@ -352,16 +386,18 @@ var TogglerDelegate = A.Component.create({
                 expanded = false;
             }
 
-            var toggler = new Toggler({
-                animated: instance.get('animated'),
-                bindDOMEvents: false,
-                bubbleTargets: [instance],
-                content: instance.findContentNode(header),
-                expanded: expanded,
-                header: header,
-                toggleEvent: instance.get('toggleEvent'),
-                transition: instance.get('transition')
-            });
+            var toggler = new Toggler(
+                {
+                    animated: animated,
+                    bindDOMEvents: false,
+                    bubbleTargets: [instance],
+                    content: instance.findContentNode(header),
+                    expanded: expanded,
+                    header: header,
+                    toggleEvent: toggleEvent,
+                    transition: transition
+                }
+            );
 
             instance.items.push(toggler);