+++ src/aui-tooltip/js/aui-tooltip-delegate.js
@@ -31,10 +31,16 @@ A.TooltipDelegate = A.Base.create('tooltip-delegate', A.Base, [], {
      * @protected
      */
     initializer: function() {
-        var instance = this;
+        var instance = this,
+            useARIA = instance.get('useARIA');
 
         instance._eventHandles = [];
+
         instance.bindUI();
+
+        if (useARIA) {
+            instance.plug(A.Plugin.Aria);
+        }
     },
 
     /**
@@ -92,6 +98,7 @@ A.TooltipDelegate = A.Base.create('tooltip-delegate', A.Base, [], {
                 opacity: instance.get('opacity'),
                 position: instance.get('position'),
                 html: instance.get('html'),
+                useARIA: instance.get('useARIA'),
                 visible: false,
                 zIndex: instance.get('zIndex')
             });
@@ -108,9 +115,15 @@ A.TooltipDelegate = A.Base.create('tooltip-delegate', A.Base, [], {
      * @protected
      */
     _onUserHideInteraction: function() {
-        var instance = this;
+        var instance = this,
+            tooltipBoundingBox = instance.getTooltip().get('boundingBox'),
+            useARIA = instance.get('useARIA');
 
         instance.getTooltip().hide();
+
+        if (useARIA) {
+            instance.aria.setAttribute('hidden', true, tooltipBoundingBox);
+        }
     },
 
     /**
@@ -122,11 +135,15 @@ A.TooltipDelegate = A.Base.create('tooltip-delegate', A.Base, [], {
      */
     _onUserShowInteraction: function(event) {
         var instance = this,
-            trigger;
-
-        trigger = event.currentTarget;
+            tooltipBoundingBox = instance.getTooltip().get('boundingBox'),
+            trigger = event.currentTarget,
+            useARIA = instance.get('useARIA');
 
         instance.getTooltip().show().set('trigger', trigger).render();
+
+        if (useARIA) {
+            instance.aria.setAttribute('hidden', false, tooltipBoundingBox);
+        }
     },
 
     /**
@@ -255,6 +272,20 @@ A.TooltipDelegate = A.Base.create('tooltip-delegate', A.Base, [], {
             validator: '_validateTriggerEvent',
             value: 'mouseenter',
             writeOnce: true
+        },
+
+        /**
+        * Boolean indicating if use of the WAI-ARIA Roles and States
+        * should be enabled.
+        *
+        * @attribute useARIA
+        * @default true
+        * @type Boolean
+        */
+        useARIA: {
+            validator: Lang.isBoolean,
+            value: true,
+            writeOnce: 'initOnly'
         },
 
         /**