+++ src/aui-tooltip/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-3106](https://issues.liferay.com/browse/AUI-3106) Make aui-tooltip accessible
 
 ## [3.1.0](https://github.com/liferay/alloy-ui/releases/tag/3.1.0)